move application approve page file location
